test: test that all generated values can be encoded
After running this test I also noticed an issue with icons, where
(inexplicably) the `blobVersionId` would sometimes be a regular string,
not a version ID. I was able to fix this by changing the definition.

I think this is a useful change on its own, but it will make future
changes (such as upgrading some dependencies) much easier.